% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control.R
\name{check_data}
\alias{check_data}
\title{Helper function to check data structure}
\usage{
check_data(
  GRADES,
  TIMES,
  TODO,
  OUTCOME,
  X,
  FIRST_YEAR,
  LAST_YEAR,
  LAST_EXAM_YEAR,
  MAX_TIME,
  LABS_EXAMS = NULL,
  LABS_OBS = NULL,
  LABS_GRADES = NULL,
  LABS_COV = NULL,
  SUBSET = NULL,
  VERBOSE = FALSE
)
}
\arguments{
\item{GRADES}{Grades matrix}

\item{TIMES}{Time matrix}

\item{TODO}{To-do matrix}

\item{OUTCOME}{Outcome vector}

\item{X}{COvariate matrix}

\item{FIRST_YEAR}{Vector containing start years}

\item{LAST_YEAR}{Vector containing last years}

\item{LAST_EXAM_YEAR}{Vector containing last exam year}

\item{MAX_TIME}{Vector containing max observable day}

\item{LABS_EXAMS}{Exam labels}

\item{LABS_OBS}{Observation labels}

\item{LABS_GRADES}{Grades labels}

\item{LABS_COV}{Covariates labels}

\item{SUBSET}{Subset of the sample to be returned}

\item{VERBOSE}{TRUE for verbose output}
}
\description{
Helper function to check data structure
}
