% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EAP.R
\name{compute_eap}
\alias{compute_eap}
\title{MAP GRTCM}
\usage{
compute_eap(FIT, TIDY = TRUE)
}
\arguments{
\item{FIT}{Output from \link{fit_EM} or \link{fit_BFGS}}

\item{TIDY}{Return tidy parameter table}
}
\description{
Compute MAP scores for Graded Response Time Censored Model
}
