% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BFGS.R
\name{fit_BFGS}
\alias{fit_BFGS}
\title{Fit measurement model and joint model with BFGS}
\usage{
fit_BFGS(DATA, GRID, WEIGHTS, THETA_START = NULL, MOD, ...)
}
\arguments{
\item{DATA}{data as per \code{\link[=check_data]{check_data()}}.}

\item{GRID}{Grid of quadrature points to be used.}

\item{WEIGHTS}{Weights for quadrature points}

\item{THETA_START}{Optional starting parameter vector}

\item{MOD}{Choose \code{full} for joint model or \code{grtcm} for measurement model.}

\item{...}{additional arguments to be passed to \code{\link[ucminf:ucminf]{ucminf::ucminf()}}}
}
\description{
Fit measurement model and joint model with BFGS
}
