% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BFGS.R
\name{fit_BFGS2}
\alias{fit_BFGS2}
\title{Fit CRGRTCM}
\usage{
fit_BFGS2(DATA, GRID, WEIGHTS, THETA_START = NULL, MOD, ...)
}
\arguments{
\item{DATA}{Dataset in long form}

\item{GRID}{Grid of quadrature points to be used.}

\item{WEIGHTS}{Weights for quadrature points}

\item{THETA_START}{Optional starting parameter vector}

\item{...}{additional arguments to be passed to \code{\link[ucminf:ucminf]{ucminf::ucminf()}}}

\item{GRADE_COL}{Label of the column where to find the grades}

\item{EXAM_COL}{Label of the column where to find exams ids}

\item{STUDID_COL}{Label of the column where to find students ids.}

\item{TIME_COL}{Label of the column where to find exams times.}

\item{MAXTIME_COL}{Label of the column where to find the maximum time observed for each student.}
}
\description{
Fit Competing Risk Graded Response Time Censored Model
}
