% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCR.R
\name{fit_CCR}
\alias{fit_CCR}
\title{Fit conditional competing risk model}
\usage{
fit_CCR(DATA, PAR_START, LATMAT)
}
\arguments{
\item{DATA}{data as per \code{\link[=check_data]{check_data()}}.}

\item{PAR_START}{Starting parameter vector}

\item{LATMAT}{Latent scores to be used for conditional competing risk model}
}
\description{
Fit conditional competing risk model
}
