% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM.R
\name{fit_EM}
\alias{fit_EM}
\title{Fit measurement model and joint model with BFGS}
\usage{
fit_EM(
  DATA,
  GRID,
  WEIGHTS,
  MOD,
  M_MAX_ITER,
  MAX_ITER,
  TOL,
  VERBOSE = TRUE,
  THETA_START = NULL
)
}
\arguments{
\item{DATA}{data as per \code{\link[=check_data]{check_data()}}.}

\item{GRID}{Grid of quadrature points to be used.}

\item{WEIGHTS}{Weights for quadrature points}

\item{MOD}{Choose \code{full} for joint model or \code{grtcm} for measurement model.}

\item{M_MAX_ITER}{Max number of iterations M-STEP.}

\item{MAX_ITER}{Maximum number of EM iterations.}

\item{TOL}{tolerance level Q-function improvement.}

\item{VERBOSE}{TRUE for verbose output.}

\item{THETA_START}{Optional starting parameter vector.}
}
\description{
Fit measurement model and joint model with BFGS
}
