% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_k_years.R
\name{predict_k_years}
\alias{predict_k_years}
\title{K-year-ahead outcome predictions}
\usage{
predict_k_years(
  FIT,
  CUTOFF_YEAR,
  HORIZON,
  YEAR_LENGTH = 365,
  GRAD_EXTENSION = 1.5
)
}
\arguments{
\item{FIT}{Fitted object returned by \code{\link[=fit_EM]{fit_EM()}} or \code{\link[=fit_BFGS]{fit_BFGS()}} with
\code{mod = "full"}.}

\item{CUTOFF_YEAR}{Integer year \code{y} whose information is used as the starting
point. Must satisfy \code{0 <= CUTOFF_YEAR}.}

\item{HORIZON}{Strictly positive integer defining how many future academic
years are evaluated. The final horizon year is \code{CUTOFF_YEAR + HORIZON}.}

\item{YEAR_LENGTH}{Number of day within each year. Censoring day is computed as \code{CUTOFF_YEAR*YEAR_LENGTH}.}

\item{GRAD_EXTENSION}{Scalar multiplier to account for extra graduation sessions. Last day useful for graduation is computed as \code{CUTOFF_YEAR*YEAR_LENGTH+ GRAD_EXTENSION*YEAR_LENGTH}.}
}
\value{
A tibble with one row per student containing cumulative probabilities
of graduation, dropout, transfer, and remaining enrolled by the end of the
target horizon.
}
\description{
Compute the probability that each student experiences graduation, dropout,
transfer, or remains enrolled by the end of academic year
\code{CUTOFF_YEAR + HORIZON}, conditioning on the data available up to the end of
\code{CUTOFF_YEAR}. Probabilities are obtained by iterating
\code{\link[=predict_one_year]{predict_one_year()}} over successive horizons and accumulating the implied
hazards.
}
