% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict_one_year}
\alias{predict_one_year}
\title{Predicted outcome probabilities one year ahead}
\usage{
predict_one_year(FIT, CUTOFF_YEAR, YEAR_LENGTH = 365, GRAD_EXTENSION = 1.5)
}
\arguments{
\item{FIT}{Fitted object returned by \code{\link[=fit_EM]{fit_EM()}} or \code{\link[=fit_BFGS]{fit_BFGS()}}.}

\item{CUTOFF_YEAR}{Consider data up to the end of \code{CUTOFF_YEAR}.}

\item{YEAR_LENGTH}{Number of day within each year. Censoring day is computed as \code{CUTOFF_YEAR*YEAR_LENGTH}.}

\item{GRAD_EXTENSION}{Scalar multiplier to account for extra graduation sessions. Last day useful for graduation is computed as \code{CUTOFF_YEAR*YEAR_LENGTH+ GRAD_EXTENSION*YEAR_LENGTH}.}
}
\description{
Predicted outcome probabilities one year ahead
}
