% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_student_progress.R
\name{predict_student_progress}
\alias{predict_student_progress}
\title{Sequential risk updates for a single student}
\usage{
predict_student_progress(
  FIT,
  SUBJECT_ID,
  horizons = NULL,
  YEAR_LENGTH = 365,
  GRAD_EXTENSION = 1.5
)
}
\arguments{
\item{FIT}{Fitted object returned by \code{\link[=fit_EM]{fit_EM()}} or \code{\link[=fit_BFGS]{fit_BFGS()}}.}

\item{SUBJECT_ID}{Character identifier of the student (must match
\code{FIT$data$labs$obs}).}

\item{horizons}{Integer vector of academic years to forecast relative to the
current cutoff. By default the function projects to every remaining year.}

\item{YEAR_LENGTH}{Number of day within each year. Censoring day is computed as \code{CUTOFF_YEAR*YEAR_LENGTH}.}

\item{GRAD_EXTENSION}{Scalar multiplier to account for extra graduation sessions. Last day useful for graduation is computed as \code{CUTOFF_YEAR*YEAR_LENGTH+ GRAD_EXTENSION*YEAR_LENGTH}.}
}
\description{
Sequential risk updates for a single student
}
