% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stud_careers}
\alias{stud_careers}
\title{Student careers}
\format{
\subsection{`stud_careers``}{

\describe{
\item{data_dims}{Gather info about data dimensions. List containing number of students (\code{n_obs}), number of exams (\code{n_exams}), number of grades categories (\code{n_grades}), number of covariates (\code{n_cov}).}
\item{par_dims}{Gather info about model dimension. List containing number of parameters releted to measurement model (\code{grtcm}), number of parameters related to outcome model (\code{cr}), number of parameters related to structural model (\code{lat}).}
\item{labs}{List containing students, exams and grades labels.}
\item{todoMat}{\code{n_obs} x \code{n_exams} matrix of logical values denoting the plan of study for each student.}
\item{gradesMat}{\code{n_obs} x \code{n_exams} matrix collecting students' grades (categories) on each exam.}
\item{timeMat}{\code{n_obs} x \code{n_exams} matrix of times (0 denotes enrollment day) before passing each exam.}
\item{outcome}{\code{n_obs}-dimensional vector collecting academic outcome information. Encoding: 0 for still enrolled; 1 for graduation, 2 for dropout; 3 for transfer.}
\item{first_year}{\code{n_obs}-dimensional vector collecting first enrollment year.}
\item{last_year}{\code{n_obs}-dimensional vector collecting last enrollment year.}
\item{yle}{\code{n_obs}-dimensional vector collecting the year of study plan completio.}
\item{max_time}{\code{n_obs}-dimensional vector collecting max observed day.}
\item{X}{\code{n_obs} x \code{n_exams} matrix of covariates.}
}
}
}
\source{
Alfonzetti, G., and Battauz, M. "A joint model for academic outcomes and students' exam performance with informative censoring".
}
\usage{
stud_careers
}
\description{
Dataset collecting information about students careers, including exams-related grades and times as well as academic outcomes.
}
\keyword{datasets}
