% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SE.R
\name{compute_stderr}
\alias{compute_stderr}
\title{Compute standard errors from GRTCM fit}
\usage{
compute_stderr(
  FIT,
  METHOD = c("sample", "num", "bfgs"),
  TIDY = TRUE,
  GRID = NULL,
  WEIGHTS = NULL,
  VERBOSE = FALSE
)
}
\arguments{
\item{FIT}{Output from \link{fit_EM} or \link{fit_BFGS}}

\item{METHOD}{Choose among \code{sample} \code{num} and \code{bfgs}.}

\item{TIDY}{Return tidy parameter table}

\item{GRID}{GH points grid}

\item{WEIGHTS}{GH weights}

\item{VERBOSE}{TRUE for verbose output}
}
\description{
Compute standard errors from GRTCM fit
}
